backend/requirements.txt
flask
flask-cors
pymongo
deepface
mtcnn
numpy
opencv-python-headless
scipy
python-dotenv
flask-bcryptz